Crate merino_suggest_providers

All of the merino suggestion providers, including tools to make and combine them.

Modules

maker

Tools to build providers from configuration.

providers

Merino providers that aren’t tied to any particular data source.

reconfigure

Helpers to reconfigure a provider and it’s inner providers.

Structs

ClientVariantFilterProvider

A provider that gives suggestions base

DebugProvider

A toy suggester to test the system.

FixedProvider

A suggester that always provides the same suggestion, with a configurable title.

IdMulti

A provider that aggregates suggestions from suggesters that tracks an ID per suggester (or suggester tree).

IdMultiProviderDetails

Metadata about a provider contained in [NamedMulti];

KeywordFilterProvider

A combinator provider that filters the results from the wrapped provider using a blocklist from the settings.

Multi

A provider that aggregates suggestions from multiple suggesters.

NullProvider

A basic provider that never returns any suggestions.

StealthProvider

A provider that runs inner, but doesn’t return any results.

TimeoutProvider

A combinator provider that returns an empty set of suggestions if the wrapped provider takes too long.

WikiFruit

A toy suggester to test the system.

Functions

make_provider_tree

Recursive helper to build a tree of providers.

reconfigure_provider_tree

Reconfigure providers in place.